Modified the advance problem

+++ Algorithmic Toolbox/fibonacci_last_digit.cpp
@@ -10,7 +10,7 @@ int get_fibonacci_last_digit(long long n) {
       std::vector<int> f;
       f.push_back(0);
       f.push_back(1);
-      for(int i=2; i<=n; i++){
+      for(long long i=2; i<=n; i++){
         f.push_back((f[i-1] + f[i-2]) % 10);
       }
       return f.back();
